refactor: move `parseURL` into `utils.ts`
