--- packages/fatcher/src/url/index.ts
@@ -1,1 +1,0 @@
-export * from './parseURL';
